+++ packages/core/core/src/factories/apiClientFactory/index.ts
@@ -1,0 +1,9 @@
+import apiClientFactory from './apiClientFactory';
+import apiProxyFactory from './apiProxyFactory';
+export * from './types';
+
+export {
+  apiProxyFactory,
+  apiClientFactory
+};
+
